+++ src-tauri/local-crates/fit-launcher-config/src/client/mod.rs
@@ -1,1 +1,2 @@
+pub mod cookies;
 pub mod dns;
